feature: Se agerga toastr para las notificaciones.

+++ src/app/components/products/product-dumb/product-dumb.component.ts
@@ -1,6 +1,7 @@
-import { Component, OnInit } from '@angular/core';
+import {Component, OnInit} from '@angular/core';
 import {ProductService} from '../../../services/product.service';
 import {Product} from '../../../models/product';
+import {ToastrService} from 'ngx-toastr';
 
 @Component({
   selector: 'app-product-dumb',
@@ -11,7 +12,8 @@ export class ProductDumbComponent implements OnInit {
 
   productList: Product[];
 
-  constructor(readonly productService: ProductService) { }
+  constructor(readonly productService: ProductService, readonly toastr: ToastrService) {
+  }
 
   ngOnInit() {
     this.productService.getProducts()
@@ -20,7 +22,7 @@ export class ProductDumbComponent implements OnInit {
         this.productList = [];
         item.forEach(element => {
           const product = element.payload.toJSON();
-          product["$key"] = element.key;
+          product['$key'] = element.key;
           this.productList.push(product as Product);
         });
       });
@@ -31,7 +33,10 @@ export class ProductDumbComponent implements OnInit {
   }
 
   onDelete($key: string) {
-    this.productService.deleteProduct($key);
+    if (confirm('Are you sure?')) {
+      this.productService.deleteProduct($key);
+      this.toastr.success('Successfull Operation', 'Producto eliminado', {timeOut: 1000});
+    }
   }
 
 }
